validation avant de delete projet et setProjet

+++ controleur/DeleteProjetAction.php
@@ -6,10 +6,16 @@ class DeleteProjetAction implements Action {
 	public function execute(){
         $projetDao = new ProjetDAO();
         $tacheDao = new TacheDAO();
-        $projet = $projetDao->find($_REQUEST['numDelete']);
-        if($projet!=null){
-            $projetDao->delete($projet);
-            $tacheDao->deleteByNum($projet);
+        if(isset($_REQUEST['numDelete'])){
+            $proTemp = new Projets();
+            $proTemp->setEmail($_SESSION['current_email']);
+            $proTemp->setNumProjet($_REQUEST['numDelete']);
+
+            $projet = $projetDao->findByProjet($proTemp);
+            if($projet!=null and $projet->getRole()=="admin"){
+                $projetDao->delete($projet);
+                $tacheDao->deleteByNum($projet);
+            }
         }
 		return "listeProjets";
 	}
